WIP: Operator interface
