+++ rlst/Cargo.toml
@@ -1,6 +1,3 @@
-[features]
-strict = []
-
 [package]
 name = "rlst"
 version = "0.0.1"
@@ -25,6 +22,9 @@ rlst-proc-macro = { path = "../proc-macro"}
 num = "0.4"
 approx = { version = "0.5", features=["num-complex"] }
 
+mpi = { version="0.7.0", optional=true }
+mpi-sys = { version="0.2.1", optional=true }
+
 
 [target.aarch64-apple-darwin.dependencies]
 blas-src = { version = "0.9", features = ["accelerate"]}
@@ -39,4 +39,8 @@ paste = "1"
 rustdoc-args = [ "--html-in-header", "katex-header.html" ]
 
 
-
+[features]
+strict = []
+default = ["mpi", "umfpack"]
+mpi = ["dep:mpi", "dep:mpi-sys", "rlst-sparse/mpi"]
+umfpack = ["rlst-sparse/umfpack"]
