+++ sparse/Cargo.toml
@@ -14,7 +14,7 @@ rand = "0.8"
 rand_distr = "0.4"
 approx = { version = "0.5", features=["num-complex"] }
 itertools = "0.12"
-# rlst-umfpack = { path = "../umfpack", optional = true}
+rlst-umfpack = { path = "../umfpack", optional = true}
 
 
 
@@ -27,9 +27,8 @@ rand_chacha = "0.3"
 name = "rlst_sparse"
 
 [features]
-default = ["mpi","mpi-sys"]
-mpi = ["dep:mpi"]
-# umfpack = ["dep:rlst-umfpack"]
-umfpack = []
+default = []
+mpi = ["dep:mpi", "mpi-sys"]
+umfpack = ["dep:rlst-umfpack"]
 
 strict = []
