+++ sparse/src/lib.rs
@@ -1,15 +1,12 @@
 #![cfg_attr(feature = "strict", deny(warnings))]
 
+#[cfg(feature = "mpi")]
 pub mod distributed_vector;
+
+#[cfg(feature = "mpi")]
 pub mod ghost_communicator;
+
 pub mod index_layout;
-// pub mod operator_interface;
 pub mod sparse;
 pub mod tools;
 pub mod traits;
-
-// #[cfg(feature = "mpi")]
-// pub use distributed_vector::*;
-
-// #[cfg(test)]
-// mod tests {}
