+++ sparse/src/sparse.rs
@@ -1,6 +1,9 @@
 pub mod csc_mat;
 pub mod csr_mat;
+
+#[cfg(feature = "mpi")]
 pub mod mpi_csr_mat;
+
 pub mod tools;
 #[cfg(feature = "umfpack")]
 pub mod umfpack;
