+++ sparse/src/sparse/csc_mat.rs
@@ -198,79 +198,3 @@ impl<Item: RlstScalar> rlst_dense::traits::Shape<2> for CscMatrix<Item> {
         self.shape
     }
 }
-
-#[cfg(test)]
-mod test {
-
-    use rlst_dense::traits::AijIterator;
-
-    use super::*;
-
-    #[test]
-    fn test_csc_from_aij() {
-        // Test the matrix [[1, 2], [3, 4]]
-        let rows = vec![0, 0, 1, 1, 0];
-        let cols = vec![0, 1, 0, 1, 1];
-        let data = vec![1.0, 2.0, 3.0, 4.0, 6.0];
-
-        let csc = CscMatrix::from_aij([2, 2], &rows, &cols, &data).unwrap();
-
-        assert_eq!(csc.data().len(), 4);
-        assert_eq!(csc.indices().len(), 4);
-        assert_eq!(csc.indptr().len(), 3);
-        assert_eq!(csc.data()[2], 8.0);
-
-        // Test the matrix [[0, 2.0, 0.0], [0, 0, 0], [0, 0, 0]]
-        let rows = vec![0, 2, 0];
-        let cols = vec![1, 2, 1];
-        let data = vec![2.0, 0.0, 3.0];
-
-        let csc = CscMatrix::from_aij([3, 3], &rows, &cols, &data).unwrap();
-
-        assert_eq!(csc.indptr()[0], 0);
-        assert_eq!(csc.indptr()[1], 0);
-        assert_eq!(csc.indptr()[2], 1);
-        assert_eq!(csc.indptr()[3], 1);
-        assert_eq!(csc.data()[0], 5.0);
-    }
-
-    #[test]
-    fn test_csc_matmul() {
-        // Test the matrix [[1, 2], [3, 4]]
-        let rows = vec![0, 0, 1, 1];
-        let cols = vec![0, 1, 0, 1];
-        let data = vec![1.0, 2.0, 3.0, 4.0];
-
-        let csc = CscMatrix::from_aij([2, 2], &rows, &cols, &data).unwrap();
-
-        // Execute 2 * [1, 2] + 3 * A*x with x = [3, 4];
-        // Expected result is [35, 79].
-
-        let x = vec![3.0, 4.0];
-        let mut res = vec![1.0, 2.0];
-
-        csc.matmul(3.0, &x, 2.0, &mut res);
-
-        assert_eq!(res[0], 35.0);
-        assert_eq!(res[1], 79.0);
-    }
-
-    #[test]
-    fn test_aij_iterator() {
-        let rows = vec![2, 3, 4, 4, 6];
-        let cols = vec![1, 1, 3, 3, 4];
-        let data = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-
-        // This matrix has a zero col at the beginning one in between and several zero cols at the end.
-        let csr = CscMatrix::from_aij([10, 20], &rows, &cols, &data).unwrap();
-
-        let aij_data: Vec<(usize, usize, f64)> = csr.iter_aij().collect();
-
-        assert_eq!(aij_data.len(), 4);
-
-        assert_eq!(aij_data[0], (2, 1, 1.0));
-        assert_eq!(aij_data[1], (3, 1, 2.0));
-        assert_eq!(aij_data[2], (4, 3, 7.0));
-        assert_eq!(aij_data[3], (6, 4, 5.0));
-    }
-}
