+++ sparse/src/sparse/csr_mat.rs
@@ -196,80 +196,3 @@ impl<Item: RlstScalar> Shape<2> for CsrMatrix<Item> {
         self.shape
     }
 }
-
-#[cfg(test)]
-mod test {
-
-    use rlst_dense::traits::AijIterator;
-
-    use super::*;
-
-    #[test]
-    fn test_csr_from_aij() {
-        // Test the matrix [[1, 2], [3, 4]]
-        let rows = vec![0, 0, 1, 1, 0];
-        let cols = vec![0, 1, 0, 1, 1];
-        let data = vec![1.0, 2.0, 3.0, 4.0, 6.0];
-
-        let csr = CsrMatrix::from_aij([2, 2], &rows, &cols, &data).unwrap();
-
-        assert_eq!(csr.data().len(), 4);
-        assert_eq!(csr.indices().len(), 4);
-        assert_eq!(csr.indptr().len(), 3);
-        assert_eq!(csr.data()[1], 8.0);
-
-        //Test the matrix [[0, 0, 0], [2.0, 0, 0], [0, 0, 0]]
-        let rows = vec![1, 2, 1];
-        let cols = vec![0, 2, 0];
-        let data = vec![2.0, 0.0, 3.0];
-
-        let csr = CsrMatrix::from_aij([3, 3], &rows, &cols, &data).unwrap();
-
-        assert_eq!(csr.indptr()[0], 0);
-        assert_eq!(csr.indptr()[1], 0);
-        assert_eq!(csr.indptr()[2], 1);
-        assert_eq!(csr.indptr()[3], 1);
-        assert_eq!(csr.data()[0], 5.0);
-    }
-
-    #[test]
-    fn test_csr_matmul() {
-        // Test the matrix [[1, 2], [3, 4]]
-        let rows = vec![0, 0, 1, 1];
-        let cols = vec![0, 1, 0, 1];
-        let data = vec![1.0, 2.0, 3.0, 4.0];
-
-        let csr = CsrMatrix::from_aij([2, 2], &rows, &cols, &data).unwrap();
-
-        // Execute 2 * [1, 2] + 3 * A*x with x = [3, 4];
-        // Expected result is [35, 79].
-
-        let x = vec![3.0, 4.0];
-        let mut res = vec![1.0, 2.0];
-
-        csr.matmul(3.0, &x, 2.0, &mut res);
-
-        assert_eq!(res[0], 35.0);
-        assert_eq!(res[1], 79.0);
-    }
-
-    #[test]
-    fn test_aij_iterator() {
-        let rows = vec![2, 3, 4, 4, 6];
-        let cols = vec![0, 1, 0, 2, 1];
-        let data = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-
-        // This matrix has a zero row at the beginning one in between and several zero rows at the end.
-        let csr = CsrMatrix::from_aij([10, 3], &rows, &cols, &data).unwrap();
-
-        let aij_data: Vec<(usize, usize, f64)> = csr.iter_aij().collect();
-
-        assert_eq!(aij_data.len(), 5);
-
-        assert_eq!(aij_data[0], (2, 0, 1.0));
-        assert_eq!(aij_data[1], (3, 1, 2.0));
-        assert_eq!(aij_data[2], (4, 0, 3.0));
-        assert_eq!(aij_data[3], (4, 2, 4.0));
-        assert_eq!(aij_data[4], (6, 1, 5.0));
-    }
-}
