+++ sparse/src/traits.rs
@@ -1,5 +1,3 @@
 // Interface traits
 
 pub mod index_layout;
-pub mod indexable_matrix;
-pub mod indexable_vector;
